Auto-Repair Tuning (#652)
* difficulty modifiers for repair rate and drain in config

* changed autorepair request-repair pattern to wait until a repair is fulfilled before issuing the next request; moved integration tests for efficiency

* auto-repair accounts for delay between request-reply when schduling next request; base ntu drain is halved; auto-repair mending values adjusted based on wiki times

* moving config call into auto-repair mixin

* deleting the old integration testing code

* obligatory test fixes

* more of the same

* correcting exceptions with calculating subsequent auto-repair time and transfer messaging behavior with WarpGate objects; transfer animation and transfer delivery now have slightly different timing; wrote  bunch of tests that don't work in the standard manner, but do pass

+++ src/main/scala/net/psforever/objects/serverobject/structures/AmenityDefinition.scala
@@ -7,7 +7,7 @@ import net.psforever.objects.vital._
 import net.psforever.objects.vital.resistance.ResistanceProfileMutators
 import net.psforever.objects.vital.resolution.DamageResistanceModel
 
-final case class AutoRepairStats(amount: Int, start: Long, repeat: Long, drain: Float)
+final case class AutoRepairStats(amount: Float, start: Long, repeat: Long, drain: Float)
 
 abstract class AmenityDefinition(objectId: Int)
     extends ObjectDefinition(objectId)
